RocketMedia Skills Activity - Nathaniel
Results should match what's in the PSD. Visit index.html to see it in
action.
I wasn't sure what to add as far as Javascript was concerned. However,
it seemed reasonable to make the widget disappear when the 'close'
button was clicked. So, I added that, and I made it so that anyone
clicking on the blue square with the house on it would be redirected to
another website (for good measure).
For fun, I added a little effect on the words 'Home Automation' when the
widget is hovered over, and when you hover over the blue box with the
white house, the house turns the same yellow as 'Learn More'.